Added model view module
The module implements simplified model view mechanics using a qt widget. As the
qt model/view utilities seemed to complex for the given case, a simpler model
was implemented. Several views can observe a model. If the state of the model
is changed, an update is triggered at the corresponding views. The model of a
view can be set at view construction or using the model property, hiding the
model view connection logic.