Updated directory parsing of encoder logs
The parse_directories function of the encoder log class, does not
automatically parse the 'log' subfolder anymore, as this is done
automatically by the parse_url method. Note, that as a result of
this change, the file menu entry does not parse log subfolders
automatically anymore. Automatical parsing of 'log' subfolders is
still supported for drop of files.