Added custom signal `items_changed` for efficient gui updating
Up until now, the GUI was updated using the `rowsInserted` and
`rowsRemoved` signals. The problem here was, that they were called
n-times, if n-items were added/removed to the model in the
previous implementation. As each signal call resulted in GUI
updates, this was quite inefficient. The interface methods for
changes of the model were reimplemented to allow adding/removing
a collection of item at once. The `rowsInserted` and `rowsRemoved`
functions are still called for each single element, so the views
update correctly, especially for the tree. But, custom functions
in main now use the new `items_changed` signal, which is emitted
only ones, after the whole collection has been processed.