Changed the encoder log class' data export interface
Encoder log class exports its data by a list of tuples. The tuples
contain identifiers, by which data from different encoder logs
can be associated with each other, and the tuples contain the data
itself. The identifiers can eg. be used to associate summary data
of different encoder logs with each other. The data itself is as
before a dictionary tree of variables containing lists of data
at the leafs.