Added functions to join dictionary variable trees of encoder logs
The functions parse a collection of encoder logs, and create a
joined version of their dictionary trees. The keys of the
dictionary trees are variables, the leafs are lists of values. The
lists of values are combined to plot data objects. Values from
different encoder logs at the same variable position are joined
if their identifiers are equal. If not, the plot data objects with
different identifiers are contained in parallal, eg. in a list,
at the variable position in the output tree.