Add bjontegaard ci table updates and plotting
Bjontegaard calculations are now adjusted to the mode of ci (average, worst, best), which can be adjusted in a new combo box. The plot will show the selected curves (borders) in the confidence intervals. The anchor can be selected by double clicking the table headers. Additionally, mixed plotting of ci and non-ci curves are now possible.